Removing the cards!
Unfortunately, this is not really good, as the page is not really mobile ready anymore. The cards are more flexible for sure. But as we are already using tables, and tables are also killing the mobile-version, thats ok for now. But yeah, we need to find a structure with which we can rebuild this static elements with more flexible ones. We have missed that.